+++ MANO-API/etsi-models/3.5.1/model/src/main/java/com/ubiqube/etsi/mano/nfvo/v351/model/vnf/VnfPackageSoftwareImageInfo.java
@@ -17,6 +17,7 @@
 package com.ubiqube.etsi.mano.nfvo.v351.model.vnf;
 
 import java.time.OffsetDateTime;
+import java.util.Map;
 import java.util.Objects;
 
 import javax.validation.Valid;
@@ -28,7 +29,6 @@ import org.springframework.validation.annotation.Validated;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonValue;
-import java.util.Map;
 import com.ubiqube.etsi.mano.nfvo.v351.model.vnfsnapshotpkgm.Checksum;
 
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -173,13 +173,13 @@ public class VnfPackageSoftwareImageInfo {
 	private OffsetDateTime createdAt = null;
 
 	@JsonProperty("minDisk")
-	private Integer minDisk = null;
+	private Long minDisk = null;
 
 	@JsonProperty("minRam")
-	private Integer minRam = null;
+	private Long minRam = null;
 
 	@JsonProperty("size")
-	private Integer size = null;
+	private Long size = null;
 
 	@JsonProperty("userMetadata")
 	private Map<String, String> userMetadata = null;
@@ -393,7 +393,7 @@ public class VnfPackageSoftwareImageInfo {
 		this.createdAt = createdAt;
 	}
 
-	public VnfPackageSoftwareImageInfo minDisk(final Integer minDisk) {
+	public VnfPackageSoftwareImageInfo minDisk(final Long minDisk) {
 		this.minDisk = minDisk;
 		return this;
 	}
@@ -407,15 +407,15 @@ public class VnfPackageSoftwareImageInfo {
 	@NotNull
 
 	@Min(0)
-	public Integer getMinDisk() {
+	public Long getMinDisk() {
 		return minDisk;
 	}
 
-	public void setMinDisk(final Integer minDisk) {
+	public void setMinDisk(final Long minDisk) {
 		this.minDisk = minDisk;
 	}
 
-	public VnfPackageSoftwareImageInfo minRam(final Integer minRam) {
+	public VnfPackageSoftwareImageInfo minRam(final Long minRam) {
 		this.minRam = minRam;
 		return this;
 	}
@@ -429,15 +429,15 @@ public class VnfPackageSoftwareImageInfo {
 	@NotNull
 
 	@Min(0)
-	public Integer getMinRam() {
+	public Long getMinRam() {
 		return minRam;
 	}
 
-	public void setMinRam(final Integer minRam) {
+	public void setMinRam(final Long minRam) {
 		this.minRam = minRam;
 	}
 
-	public VnfPackageSoftwareImageInfo size(final Integer size) {
+	public VnfPackageSoftwareImageInfo size(final Long size) {
 		this.size = size;
 		return this;
 	}
@@ -451,11 +451,11 @@ public class VnfPackageSoftwareImageInfo {
 	@NotNull
 
 	@Min(0)
-	public Integer getSize() {
+	public Long getSize() {
 		return size;
 	}
 
-	public void setSize(final Integer size) {
+	public void setSize(final Long size) {
 		this.size = size;
 	}
